Fix update ast.Type by reference
Both ast.Value.ExpectedType and ast.VariableDefinition.Type are the pointers,
and it is easy to update values by accidence.

Therefore creating a copy, so changes in one validation rule (E.g. VariablesInAllowedPosition)
would not affect the behavior of the other rule (ProvidedRequiredArguments).

Not doing deep copy, so change should not have a big impact for memory or CPU even for large GraphQL applications.

Testing private methods, to not be confused by complex golang syntax.
(those are pure functions, so in case testing package is changed – refactoring should not be hard)