fix(validator): allow nullable variables for nonnull args with default
The validation was failing for arguments that are defined as required, but have
a default value, preventing a non-required argument to override their value,
while it should work as expected.

Given the tests are imported from JS, I'm not sure how to add a new one there.
Here's an example schema and query to reproduce this:
```
type Query {
  user(username: String! = "foo"): User
}

type User {
  username: String!
}
```

```
query Foo($username: String) {
  user(username: $username) {
    username
  }
}
```